requirements-gpu.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.0
easydict==1.13
faiss-cpu==1.12.0
filelock==3.19.1
filterpy==1.4.5
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lapx==0.9.2
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.19.1
onnxruntime-gpu==1.23.2
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
prettytable==3.16.0
protobuf==4.25.8
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
simsimd==6.5.3
six==1.17.0
sounddevice==0.5.3
stringzilla==4.2.3
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
# PyTorch with CUDA 12.4 support
# Note: These versions require installation from PyTorch's index, not PyPI 
# Install with: pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
# For now using standard versions that work from PyPI (will auto-detect CUDA if available)
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.226
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
